fix(pump): fix curve management
